Correct five maneuvers the caller checker caught
Mwabble afterburns for twenty ticks; it was coasting at full speed
instead.  Mzip_past afterburns for three when the target is beyond 3000
and roughly ahead, and reaches Mtail_fire through an else rather than an
early return.  Mtarget_laser is slot 8 of the unsupported-maneuver stub,
not a second entry into Mbest_strafe.  Msit_n_spin's sequence step 1 only
steadies the ship -- the close_behind test that reset the maneuver and
jumped to step 10 was never there.

close_behind returns a short, so callers sign-extend the result; that
also brings Mzip_past to an exact match.  steer_away_from_object returns
void.  Restore the non-intrinsic abs in Msit_n_spin and smart.c the way
auto.c and disk.c already do.

+++ src/logic.c
@@ -1880,7 +1880,7 @@ void ComputePointBehindObject(short obj, unsigned short distance,
 }
 
 /* Function start: 0x42A876 */
-unsigned int close_behind(short range)
+short close_behind(short range)
 {
     if (g_nTargetRange_0049319c < range &&
         g_nTargetFacing_00493198 < -0x32)
